+++ src/Aks/Aks/ChangeLog.md
@@ -22,6 +22,7 @@
 * Added parameter `-EnableUltraSSD` for `New-AzAksCluster` and `New-AzAksNodePool`
 * Added parameter `-NodeKubeletConfig` for `New-AzAksCluster`, `-KubeletConfig` for `New-AzAksNodePool`
 * Added parameter `-NodeLinuxOSConfig` for `New-AzAksCluster`, `-LinuxOSConfig` and `New-AzAksNodePool`
+* Added parameter `-NodeMaxSurge` for `New-AzAksCluster`, `-MaxSurge` for `New-AzAksNodePool` and `Update-AzAksNodePool`
 
 ## Version 5.1.0
 * Bumped API version to 2022-09-01